requirements.txt
Django>=5.1.6
dnspython>=2.4.0
aiohttp>=3.8.5
builtwith>=1.3.4
requests>=2.31.0
packaging>=23.0
httpx>=0.24.0
beautifulsoup4>=4.12.2
reportlab>=4.0.4
python-dotenv>=1.0.0
